Add a bunch of release engineering support.
deploy.cfg support for different authentication and full control over operating params.
Support for creating unique config files for the multiple awe clients to be run on a machine.
Create a monit config file to maintain a set of clients.

The latter are driven by the n-awe-clients parameter in the deploy.cfg.